Correct supportStackTracesSourceLocations capability
Currently we can return true in supportStackTracesSourceLocations even when source locations won't work. We already have a nice env-specific information on wheter it will or will not work in preprocessor defined for acceptance tests. This commit moves these defines to main Oakum library and starts using it to calculate supportStackTracesSourceLocations.

Acceptance tests can stop depending on these defines, but rather use the capability field to skip or not skip.

Unit tests also have to be enhanced with skipping logic, because now we may be returning supportStackTracesSourceLocations=false more often, depending on OS and compiler.